api/requirements.txt
fastapi==0.89.1
python-dotenv==0.21.0
requests==2.28.2